split ModInt
